4.1-rag-with-metadata/requirements.txt
chromadb==1.5.1
google.generativeai==0.8.6
